refactor: create ExecutionContext from config
